#3399 fixed handling of long integers in SWI

+++ krLanguages/swiprolog/src/main/java/swiprolog/language/PrologExpression.java
@@ -158,7 +158,7 @@ public abstract class PrologExpression implements Expression {
 		}
 
 		if (this.term.isInteger()) {
-			return Integer.toString(this.term.intValue());
+			return Long.toString(this.term.longValue());
 		}
 
 		if (this.term.isFloat()) {
@@ -191,7 +191,7 @@ public abstract class PrologExpression implements Expression {
 			case XFY:
 			case YFX:
 				return maybeBracketed(1) + " " + this.term.name() + " "
-				+ maybeBracketed(2);
+						+ maybeBracketed(2);
 			case XF:
 				return maybeBracketed(1) + " " + this.term.name() + " ";
 			default:
